ispravljeno kreiranje porudzbine

+++ NaruciOnline/src/model/Kupac.java
@@ -3,7 +3,7 @@ package model;
 public class Kupac extends Korisnik {
 	public TipKupca tipKupca;
 	public Korpa korpa;
-	public Integer brojBodova;
+	public double brojBodova;
 	public String svePorudzbine;
 	
 	public Kupac() {
